quiz_project/quiz_app/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers
from .models import Question

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['username', 'email']

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = "__all__"
